TUI 2.3.0b8
The timestamp for entries in the Log window now includes the date.
Corrected units of distance in Slew window from parsecs to au (thanks to Russet for noticing).
Corrected file names of built-in grids (az->alt) (thanks to Russet for noticing).
In the Status window the instrument name is red if the instrument is "?" (unknown).