TUI 2.5.0b5
Improved ARCTIC's filter wheel display. Now if a move ends
with cmdFilter != filterName, the current filter is shown
in pink, the user filter is set to the commanded filter
and you can command a move to any filter position.
Also improved the ARCTIC test data by adding some missing
data and removing keywords that are not output, and adding
an animated state that shows a move where the filter did not
go to the commanded position.